refactor(auth): split entra code into smaller modules
