+++ auth/utils/entra-auth/token.js
@@ -1,45 +1,10 @@
 import { createPublicKey } from 'node:crypto';
 import got from 'got';
 import jwt from 'jsonwebtoken';
+import { getEntraConfig, getEntraIssuer, getEntraJwksUrl, getEntraRedirectUri } from './config.js';
 
-const DEFAULT_ENTRA_SCOPE = 'openid profile email';
 const DEFAULT_ENTRA_JWKS_CACHE_TTL_MS = 60 * 1000;
-const TRUE_VALUES = ['1', 'true', 'yes', 'on'];
 const entraJwksCacheByTenant = new Map();
-
-/**
- * Returns Entra configuration from environment variables.
- *
- * @returns {{ clientId: string | undefined, clientSecret: string | undefined, tenantId: string | undefined, scope: string }}
- */
-export function getEntraConfig() {
-    return {
-        clientId: process.env.ENTRA_CLIENT_ID,
-        clientSecret: process.env.ENTRA_CLIENT_SECRET,
-        tenantId: process.env.ENTRA_TENANT_ID,
-        scope: process.env.ENTRA_SCOPE || DEFAULT_ENTRA_SCOPE
-    };
-}
-
-/**
- * Builds the expected issuer URL for Entra v2.0 tokens.
- *
- * @param {string} tenantId - Entra tenant identifier.
- * @returns {string}
- */
-function getEntraIssuer(tenantId) {
-    return `https://login.microsoftonline.com/${tenantId}/v2.0`;
-}
-
-/**
- * Builds the Entra JWKS URL used to resolve token signing keys.
- *
- * @param {string} tenantId - Entra tenant identifier.
- * @returns {string}
- */
-function getEntraJwksUrl(tenantId) {
-    return `https://login.microsoftonline.com/${tenantId}/discovery/v2.0/keys`;
-}
 
 /**
  * Returns JWKS cache TTL in milliseconds.
@@ -150,97 +115,6 @@ async function getEntraSigningKey(idToken, tenantId) {
 }
 
 /**
- * Determines if interactive fallback is enabled after silent SSO fails.
- *
- * Defaults to enabled when ENTRA_INTERACTIVE_FALLBACK is unset.
- *
- * @returns {boolean}
- */
-export function isEntraInteractiveFallbackEnabled() {
-    const rawValue = process.env.ENTRA_INTERACTIVE_FALLBACK;
-
-    if (rawValue == null) {
-        return true;
-    }
-
-    return TRUE_VALUES.includes(String(rawValue).toLowerCase());
-}
-
-/**
- * Determines whether all required Entra settings are present.
- *
- * @returns {boolean}
- */
-export function isEntraConfigured() {
-    const { clientId, clientSecret, tenantId } = getEntraConfig();
-    return Boolean(clientId && clientSecret && tenantId);
-}
-
-/**
- * Builds the callback URI using trusted configuration.
- *
- * In production this requires APP_BASE_URL to avoid relying on request Host headers.
- * In non-production, request protocol/host fallback is allowed for local development.
- *
- * @param {import('express').Request} req - Express request used to resolve protocol and host.
- * @returns {string}
- */
-export function getEntraRedirectUri(req) {
-    const appBaseUrl = process.env.APP_BASE_URL;
-
-    if (typeof appBaseUrl === 'string' && appBaseUrl.trim().length > 0) {
-        return `${appBaseUrl.replace(/\/+$/, '')}/auth/callback`;
-    }
-
-    if (process.env.NODE_ENV === 'production') {
-        throw new Error('APP_BASE_URL must be set in production for Entra redirect URI');
-    }
-
-    const protocol = req.protocol;
-    const host = req.get('host');
-    return `${protocol}://${host}/auth/callback`;
-}
-
-/**
- * Builds an Entra authorize URL for the authorization code flow.
- *
- * @param {import('express').Request} req - Express request used to derive callback URI.
- * @param {string} state - OIDC state value bound to the current auth transaction.
- * @param {string} nonce - OIDC nonce value validated against the returned id token.
- * @param {{ prompt?: string, loginHint?: string, domainHint?: string }} [options] - Optional authorize request parameters.
- * @returns {string}
- */
-export function buildEntraAuthorizeUrl(req, state, nonce, options = {}) {
-    const { clientId, tenantId, scope } = getEntraConfig();
-    const redirectUri = getEntraRedirectUri(req);
-    const authUrl = `https://login.microsoftonline.com/${tenantId}/oauth2/v2.0/authorize`;
-
-    const params = new URLSearchParams({
-        client_id: clientId,
-        response_type: 'code',
-        redirect_uri: redirectUri,
-        response_mode: 'query',
-        scope,
-        state,
-        nonce
-    });
-
-    if (options.prompt) {
-        params.set('prompt', options.prompt);
-    }
-
-    if (options.loginHint) {
-        params.set('login_hint', options.loginHint);
-    }
-
-    if (options.domainHint) {
-        params.set('domain_hint', options.domainHint);
-    }
-
-    return `${authUrl}?${params.toString()}`;
-}
-
-/**
  * Exchanges an authorization code for tokens at Entra token endpoint.
  *
  * @param {import('express').Request} req - Express request used to derive callback URI.
@@ -306,16 +180,6 @@ export async function decodeAndValidateEntraIdToken(idToken, expectedNonce) {
     }
 
     return claims;
-}
-
-/**
- * Extracts a stable username from Entra claims.
- *
- * @param {Record<string, any>} claims - Decoded id token claims.
- * @returns {string}
- */
-export function getUsernameFromEntraClaims(claims) {
-    return claims.preferred_username || claims.email || claims.upn || claims.sub || 'entra-user';
 }
 
 /**
